requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
loguru==0.7.2
pyinstaller==6.4.0
pillow==10.2.0
requests==2.31.0
urllib3==2.2.1
pyyaml==6.0.1
sqlite3worker==1.1.7
psutil==5.9.5 